Refactor TDP activites search for Elasticsearch 7
In switching TDP's search from haystack to django-elasticsearch-dsl,
we also wanted to improve the search page's efficiency by avoiding
hundreds of database and elasticsearch queries per search.

This PR adds a db table to store a cache of card and facet setups.
