+++ cfgov/teachers_digital_platform/models/__init__.py
@@ -1,4 +1,7 @@
 # flake8: noqa F401
+from teachers_digital_platform.models.activity_index_page import (
+    ActivityCardMap, ActivityIndexPage, get_activity_card_map
+)
 from teachers_digital_platform.models.django import (
     ActivityAgeRange, ActivityBloomsTaxonomyLevel, ActivityBuildingBlock,
     ActivityCouncilForEconEd, ActivityDuration, ActivityGradeLevel,
@@ -6,6 +9,4 @@ from teachers_digital_platform.models.django import (
     ActivityStudentCharacteristics, ActivityTeachingStrategy, ActivityTopic,
     ActivityType, BaseActivityTaxonomy
 )
-from teachers_digital_platform.models.pages import (
-    ActivityIndexPage, ActivityPage
-)
+from teachers_digital_platform.models.pages import ActivityPage
