+++ cfgov/teachers_digital_platform/search_indexes.py
@@ -1,6 +1,6 @@
 from haystack import indexes
 
-from teachers_digital_platform.models import ActivityIndexPage, ActivityPage
+from teachers_digital_platform.models import ActivityPage
 
 
 class ActivityPageIndex(indexes.SearchIndex, indexes.Indexable):
@@ -109,7 +109,4 @@ class ActivityPageIndex(indexes.SearchIndex, indexes.Indexable):
         return ActivityPage
 
     def index_queryset(self, using=None):
-        """Only index descendants of the Activity Search page"""
-        # This is safe because ActivityIndexPage is a singleton
-        search_page = ActivityIndexPage.objects.get()
-        return self.get_model().objects.live().descendant_of(search_page)
+        return self.get_model().objects.filter(live=True)
